feature test

+++ app/Http/Controllers/Api/AnnouncementController.php
@@ -3,85 +3,38 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreAnnouncementRequest;
+use App\Http\Requests\UpdateAnnouncementRequest;
 use App\Models\Announcement;
-use Illuminate\Http\Request;
-use Illuminate\Http\Response;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Validation\ValidationException;
-
-use function PHPSTORM_META\type;
 
 class AnnouncementController extends Controller
 {
-    public function store(Request $request)
+    public function store(StoreAnnouncementRequest $request)
     {
-
-
-
-        $request->validate([
-            'title' =>  ['required', 'string'],
-            'description' => ['required', 'string'],
-            'type' =>  ['required', 'numeric'],
-            'location' => ['required', 'string'],
-            'price' => ['required', 'string'],
-            'status' => ['required', 'numeric'],
-            'student_count' => ['required', 'numeric'],
-        ]);
-
-
-        if (!in_array($request->type, [0, 1]) || !in_array($request->status, [0, 1])) {
-            return response()->json(['error' => 'Resource not found'], 404);
-        }
-
-
-        Announcement::create([
-            'user_id' => auth()->id(),
-            'title' =>  $request->title,
-            'description' =>  $request->description,
-            'type' =>  $request->type,
-            'location' =>  $request->location,
-            'price' =>  $request->price,
-            'status' =>  $request->status,
-            'student_count' =>  $request->student_count,
-        ]);
-
-        //php artisan storage:link
-
-        return response()->noContent();
+        return response()->json(auth()->user()->announcements()->create($request->validated()));
     }
 
-    public function update(Request $request, Announcement $announcement)
+    public function update(UpdateAnnouncementRequest $request, Announcement $announcement)
     {
-        $request->validate([
-            'title' =>  ['string'],
-            'description' => ['string'],
-            'type' =>  ['numeric'],
-            'location' => ['string'],
-            'price' => ['string'],
-            'status' => ['numeric'],
-            'student_count' => ['numeric'],
-        ]);
+        $validated = $request->safe()->all();
 
-        if ($request->filled('title')) {
-            $requestData['title'] = $request->title;
+        if ($request->type === 'online') {
+            $validated['location'] = null;
         }
 
-        $announcement->update([
-            'title' => $request->filled('title') ? $request->title : DB::raw('title'),
-            'description' => $request->filled('description') ? $request->description : DB::raw('description'),
-            'type' => $request->filled('type') ? $request->type : DB::raw('type'),
-            'location' => $request->filled('location') ? $request->location : DB::raw('location'),
-            'price' => $request->filled('price') ? $request->price : DB::raw('price'),
-            'status' => $request->filled('status') ? $request->status : DB::raw('status'),
-            'student_count' => $request->filled('student_count') ? $request->student_count : DB::raw('student_count'),
-        ]);
+        $announcement->update($validated);
 
-        return response()->noContent();
+        return response()->json($announcement->fresh());
     }
 
     public function destroy(Announcement $announcement)
     {
+        if (! ($announcement->user_id == auth()->id())) {
+            return abort(403);
+        }
+
         $announcement->delete();
+
         return response()->noContent();
     }
 }
